Use Text Instead Of CSS Class For Adding X & O
Use the text of X & O instead of adding an CSS class as this would be a
better way of finding out if the user has won be using ids for each of
the columns. Since using jQuery Selectors will only show how many X or O
are on the page and not the location of them.